// src/entities/process/model/types.ts
export interface Process {
  id: number
  factoryId: number
  code: string
  name: string
  description: string | null
  position?: string | null
  imageWidth?: number | null
  imageHeight?: number | null
  hasImage?: boolean
  createdAt?: string
  updatedAt?: string
}

export interface ProcessCreateRequest {
  factoryId: number
  code: string
  name: string
  description: string
  position?: string | null
  imageWidth?: number | null
  imageHeight?: number | null
}